requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.0
networkx==3.1
matplotlib==3.8.0
tqdm==4.66.1
eth-abi==3.0.0
web3==6.2.0
